reorganize quantization files
