+++ onnxruntime/contrib_ops/cpu/quantization/qlinear_global_average_pool.h
