+++ onnxruntime/core/providers/cpu/quantization/dynamicquantizelinear.cc
