+++ onnxruntime/core/providers/cpu/quantization/dynamicquantizelinear.h
